# Specialize iterator_source for random access