Optimizing this case allows the compiler to achieve a large speedup
especially in the case where the underlying range is contiguous in
memory.

The effective result is for the performance to be equivalent to directly
using `boost::iostreams::array_source<char_type>` if the source range so
allows.

This patch arose from benchmarking several varying approaches to parsing
posix timestamps here:

 - https://stackoverflow.com/questions/49693376/why-does-using-a-stream-in-boost-spirit-penalize-performance-so-much
 - https://stackoverflow.com/questions/49671674/use-a-trait-for-parsing-a-date-in-boostspirit

I'll refer to that location for some benchmark results.