@Kojoley 2018-04-07
Have you tried to replace the loop with `std::copy`?
IIRC `std::copy` must work on all iterator types and should be optimized for random iterators.